made ColorUtils and moved static functions of Color
